+++ src/core_modules/capture-core/components/Widget/widget.types.js
@@ -1,10 +1,15 @@
 // @flow
+import type { WidgetCollapsibleProps } from './widgetCollapsible.types';
+import type { WidgetNonCollapsibleProps } from './widgetNonCollapsible.types';
 
-export type Props = {|
-    header: React$Node,
-    children: React$Node,
-    open: boolean,
-    onOpen: () => void,
-    onClose: () => void,
-    ...CssClasses,
+type CollapsibleProps = {|
+    ...WidgetCollapsibleProps,
+    collapsible?: true,
 |};
+
+type NonCollapsibleProps = {|
+    ...WidgetNonCollapsibleProps,
+    collapsible: false,
+|};
+
+export type Props = CollapsibleProps | NonCollapsibleProps;
