+++ src/core_modules/capture-core/components/Widget/widgetCollapsible.types.js
@@ -1,0 +1,15 @@
+// @flow
+import type { Node } from 'react';
+
+export type WidgetCollapsibleProps = {|
+    header: Node,
+    children: Node,
+    open: boolean,
+    onOpen: () => void,
+    onClose: () => void,
+|};
+
+export type WidgetCollapsiblePropsPlain = {|
+    ...WidgetCollapsibleProps,
+    ...CssClasses,
+|};
