+++ src/core_modules/capture-core/components/Widget/widgetNonCollapsible.types.js
@@ -1,0 +1,12 @@
+// @flow
+import type { Node } from 'react';
+
+export type WidgetNonCollapsibleProps = {|
+    header: Node,
+    children: Node,
+|};
+
+export type WidgetNonCollapsiblePropsPlain = {|
+    ...WidgetNonCollapsibleProps,
+    ...CssClasses,
+|};
